Holy fudge.
First, removed bpel4ws taglib until things settle a tad more.  The
idiomatic taglib is sufficient for baseline compatibility testing
for bpel and bpml constructs.

Secondly, merged the message-initiator and message-correlator logic
so that a message-type may either be correlatored or initiatored,
but not both.

Thirdly, removed the explicit list of message-initiators on the
process-definition and syntax.  Instead, added an @initiation
attribute to process, which may contain the value 'call'
or 'message'.  If 'call', then no message initiators exist.
If 'message', then the messages from the first reachable
transitions from (in) are used as the initiating messages.