Added (pick|choice) idiom which is sorta like (switch|case) except that each branch does not explicitly have an expression guarding it.
This is mostly useful for mesage-initiated processes that can be
initiated by multiple message types.

ie:

<process initiation="message">
    <pick>
        <choice>
            <sequence>
                <receive type="foo.msg">
                    ....
                </receive>
                ...
            </sequence>
        </choice>
        <choice>
            <sequence>
                <receive type="bar.msg">
                    ....
                </receive>
                ...
            </sequence>
        </choice>
    </pick>
</process>

This process can be initiated by either a 'foo.msg' or a 'bar.msg'.

The body of the <receive> may be an atomic <action> to receive the
message and copy parts into the case attributes.

For BPEL4WS semantics, <receive> and <on-message> have an implicit
atomic action that copies the entire message to a particular
'container' in the case attrs, for example.